Change CLI to use cred instead of credential
