+++ qpc/cred/list.py
@@ -9,29 +9,28 @@
 # along with this software; if not, see
 # https://www.gnu.org/licenses/gpl-3.0.txt.
 #
-"""CredentialShowCommand is used to show a specific credential."""
+"""CredListCommand is used to list authentication credentials."""
 
 from __future__ import print_function
-import sys
 from requests import codes
 from qpc.utils import pretty_print
 from qpc.clicommand import CliCommand
-import qpc.credential as credential
+import qpc.cred as credential
 from qpc.request import GET
 from qpc.translation import _
 import qpc.messages as messages
 
 
 # pylint: disable=too-few-public-methods
-class CredentialShowCommand(CliCommand):
-    """Defines the show command.
+class CredListCommand(CliCommand):
+    """Defines the list command.
 
-    This command is for showing a credential which can
-    be associated with sources to gather facts.
+    This command is for listing credentials which can be later associated with
+    sources to gather facts.
     """
 
     SUBCOMMAND = credential.SUBCOMMAND
-    ACTION = credential.SHOW
+    ACTION = credential.LIST
 
     def __init__(self, subparsers):
         """Create command."""
@@ -39,19 +38,11 @@ class CredentialShowCommand(CliCommand):
         CliCommand.__init__(self, self.SUBCOMMAND, self.ACTION,
                             subparsers.add_parser(self.ACTION), GET,
                             credential.CREDENTIAL_URI, [codes.ok])
-        self.parser.add_argument('--name', dest='name', metavar='NAME',
-                                 help=_(messages.CRED_NAME_HELP),
-                                 required=True)
-
-    def _build_req_params(self):
-        self.req_params = {'name': self.args.name}
 
     def _handle_response_success(self):
         json_data = self.response.json()
-        if len(json_data) == 1:
-            cred_entry = json_data[0]
-            data = pretty_print(cred_entry)
-            print(data)
+        if json_data == []:
+            print(_(messages.CRED_LIST_NO_CREDS))
         else:
-            print(_(messages.CRED_DOES_NOT_EXIST % self.args.name))
-            sys.exit(1)
+            data = pretty_print(json_data)
+            print(data)
