+++ qpc/cred/tests_list.py
@@ -18,8 +18,8 @@ import requests
 import requests_mock
 from qpc.tests_utilities import HushUpStderr, redirect_stdout
 from qpc.request import CONNECTION_ERROR_MSG, SSL_ERROR_MSG
-from qpc.credential import CREDENTIAL_URI
-from qpc.credential.show import CredentialShowCommand
+from qpc.cred import CREDENTIAL_URI
+from qpc.cred.list import CredListCommand
 from qpc.utils import get_server_location, write_server_config
 
 PARSER = ArgumentParser()
@@ -29,8 +29,8 @@ write_server_config({'host': '127.0.0.1', 'port': 8000})
 BASE_URL = get_server_location()
 
 
-class CredentialShowCliTests(unittest.TestCase):
-    """Class for testing the credential show commands for qpc."""
+class CredentialListCliTests(unittest.TestCase):
+    """Class for testing the credential list commands for qpc."""
 
     def setUp(self):
         """Create test setup."""
@@ -44,73 +44,72 @@ class CredentialShowCliTests(unittest.TestCase):
         # Restore stderr
         sys.stderr = self.orig_stderr
 
-    def test_show_cred_ssl_err(self):
-        """Testing the show credential command with a connection error."""
+    def test_list_cred_ssl_err(self):
+        """Testing the list credential command with a connection error."""
         cred_out = StringIO()
-        url = BASE_URL + CREDENTIAL_URI + '?name=credential1'
+        url = BASE_URL + CREDENTIAL_URI
         with requests_mock.Mocker() as mocker:
             mocker.get(url, exc=requests.exceptions.SSLError)
-            asc = CredentialShowCommand(SUBPARSER)
-            args = Namespace(name='credential1')
+            alc = CredListCommand(SUBPARSER)
+            args = Namespace()
             with self.assertRaises(SystemExit):
                 with redirect_stdout(cred_out):
-                    asc.main(args)
+                    alc.main(args)
                     self.assertEqual(cred_out.getvalue(), SSL_ERROR_MSG)
 
-    def test_show_cred_conn_err(self):
-        """Testing the show credential command with a connection error."""
+    def test_list_cred_conn_err(self):
+        """Testing the list credential command with a connection error."""
         cred_out = StringIO()
-        url = BASE_URL + CREDENTIAL_URI + '?name=credential1'
+        url = BASE_URL + CREDENTIAL_URI
         with requests_mock.Mocker() as mocker:
             mocker.get(url, exc=requests.exceptions.ConnectTimeout)
-            asc = CredentialShowCommand(SUBPARSER)
-            args = Namespace(name='credential1')
+            alc = CredListCommand(SUBPARSER)
+            args = Namespace()
             with self.assertRaises(SystemExit):
                 with redirect_stdout(cred_out):
-                    asc.main(args)
+                    alc.main(args)
                     self.assertEqual(cred_out.getvalue(), CONNECTION_ERROR_MSG)
 
-    def test_show_cred_internal_err(self):
-        """Testing the show credential command with an internal error."""
+    def test_list_cred_internal_err(self):
+        """Testing the list credential command with an internal error."""
         cred_out = StringIO()
-        url = BASE_URL + CREDENTIAL_URI + '?name=credential1'
+        url = BASE_URL + CREDENTIAL_URI
         with requests_mock.Mocker() as mocker:
             mocker.get(url, status_code=500, json={'error': ['Server Error']})
-            asc = CredentialShowCommand(SUBPARSER)
-            args = Namespace(name='credential1')
+            alc = CredListCommand(SUBPARSER)
+            args = Namespace()
             with self.assertRaises(SystemExit):
                 with redirect_stdout(cred_out):
-                    asc.main(args)
+                    alc.main(args)
                     self.assertEqual(cred_out.getvalue(), 'Server Error')
 
-    def test_show_cred_empty(self):
-        """Testing the show credential command successfully with empty data."""
+    def test_list_cred_empty(self):
+        """Testing the list credential command successfully with empty data."""
         cred_out = StringIO()
-        url = BASE_URL + CREDENTIAL_URI + '?name=cred1'
+        url = BASE_URL + CREDENTIAL_URI
         with requests_mock.Mocker() as mocker:
             mocker.get(url, status_code=200, json=[])
-            asc = CredentialShowCommand(SUBPARSER)
-            args = Namespace(name='cred1')
-            with self.assertRaises(SystemExit):
-                with redirect_stdout(cred_out):
-                    asc.main(args)
-                    self.assertEqual(cred_out.getvalue(),
-                                     'credential "cred1" does not exist\n')
+            alc = CredListCommand(SUBPARSER)
+            args = Namespace()
+            with redirect_stdout(cred_out):
+                alc.main(args)
+                self.assertEqual(cred_out.getvalue(),
+                                 'No credentials exist yet.\n')
 
-    def test_show_cred_data(self):
-        """Testing the show credential command with stubbed data."""
+    def test_list_cred_data(self):
+        """Testing the list credential command with stubbed data."""
         cred_out = StringIO()
-        url = BASE_URL + CREDENTIAL_URI + '?name=cred1'
+        url = BASE_URL + CREDENTIAL_URI
         credential_entry = {'id': 1, 'name': 'cred1', 'username': 'root',
                             'password': '********'}
         data = [credential_entry]
         with requests_mock.Mocker() as mocker:
             mocker.get(url, status_code=200, json=data)
-            asc = CredentialShowCommand(SUBPARSER)
-            args = Namespace(name='cred1')
+            alc = CredListCommand(SUBPARSER)
+            args = Namespace()
             with redirect_stdout(cred_out):
-                asc.main(args)
-                expected = '{"id":1,"name":"cred1","password":"********",' \
-                    '"username":"root"}'
+                alc.main(args)
+                expected = '[{"id":1,"name":"cred1","password":"********",' \
+                    '"username":"root"}]'
                 self.assertEqual(cred_out.getvalue().replace('\n', '')
                                  .replace(' ', '').strip(), expected)
