+++ qpc/cred/utils.py
